pyproject.toml
[project]
name = "email-generation-assistant"
version = "0.1.0"
description = "AI-powered Email Generation Assistant with custom evaluation metrics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.11.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "sentence-transformers>=2.7.0",
    "language-tool-python>=2.8.0",
    "jsonschema>=4.21.0",
    "scikit-learn>=1.4.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "pylint>=3.0.0",
    "ipykernel>=6.29.0",
]
ui = [
    "streamlit>=1.31.0",
    "plotly>=5.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.pylint]
max-line-length = 100
